requirements.txt
cymysql==0.9.14
DateTime==4.3
Django==2.2.7
django-cymysql==2.2.0
djangorestframework==3.10.3
List==1.3.0
psycopg2==2.8.4
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.0
zope.interface==4.7.1

